chore: conditional signOffSubmit tasks
